requirements.txt
fastapi 
uvicorn 
langchain 
openai 
llama-index 
chromadb 
weaviate-client
langchain-community 
langchainhub 
python-dotenv
docx2txt
python-docx
requests
streamlit
pdfplumber
python-docx
chromadb
jinja2
tiktoken
structlog

